split hooks to pages
