+++ playground/index.html
@@ -7,7 +7,7 @@
     <link href="/css/okta-sign-in.css" type="text/css" rel="stylesheet" />
     <link href="/css/okta-plugin-a11y.css" type="text/css" rel="stylesheet" />
     <link href="/css/fonts.css" type="text/css" rel="stylesheet" />
-    <link href="/hooks/customize.css" type="text/css" rel="stylesheet" />
+    <link href="/hooks/css/customize.css" type="text/css" rel="stylesheet" />
   </head>
   
   <body>
